# openapi.yml
openapi: 3.0.0
info:
  description: Mozilla's IAM group API
  version: 1.0.0-oas3
  title: Groups
  contact:
    email: fiji@mozilla.com
  license:
    name: MPL 2.0
    url: https://www.mozilla.org/media/MPL/2.0/index.txt
paths:
  "/groups/api/v1/groups":
    post:
      summary: creates a group
      operationId: addGroup
      description: creates a new acces group
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewGroup"
        description: Inventory item to add
      responses:
        "201":
          description: group created
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
    get:
      summary: paginated groups
      description: |
        get paginated list of a groups
      parameters:
        - in: query
          name: f
          description: filter results based on f contained in group name
          required: false
          schema:
            type: string
        - in: query
          name: by
          description: order by
          required: false
          schema:
            type: string
            enum: ["MemberCount", "NameAsc", "NameDesc"]
        - in: query
          name: "n"
          description: next page id (offset)
          required: false
          schema:
            type: integer
            default: 0
        - in: query
          name: s
          description: size of the result
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: a page of members
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: "#/components/schemas/DisplayMember"
                  next:
                    type: integer
                    example: 20
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
  "/groups/api/v1/groups/{groupName}":
    get:
      summary: group information
      operationId: viewGroup
      description: |
        Retrieve information about a given group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
      responses:
        "200":
          description: the group information requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
    put:
      summary: updates a group
      description: |
        Update information and settings for a given group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
        description: Updates for a existing group
      responses:
        "201":
          description: updates applied
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
    delete:
      summary: deletes a group
      description: |
        Deletes a group and all related data
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
      responses:
        "201":
          description: group deleted
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
  "/groups/api/v1/groups/{groupName}/details":
    get:
      summary: detailed group information
      description: Detailed information for group view
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
      responses:
        "200":
          description: a page of members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetails"
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
...
  "/groups/api/v1/invitations/{groupName}":
    get:
      summary: group invitations
      description: get invitations for a group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
      responses:
        "200":
          description: list of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
    post:
      summary: invite a member
      description: invite a member to a group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_uuid:
                  type: string
                  format: uuid
                  example: 34a33a8e-d035-4051-acac-497a79235e73
                invitation_expiration:
                  description: expiration period in days
                  type: integer
                  example: 5
                group_expiration:
                  description: expiration period in days
                  type: integer
                  example: 180
      responses:
        "201":
          description: member invited
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
  "/groups/api/v1/invitations/{groupName}/{userUuid}":
    put:
      summary: update an invitation
      description: update inviation of a user
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
        - in: path
          name: userUuid
          description: user uuid of invitation to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                invitation_expiration:
                  description: expiration period in days
                  type: integer
                  example: 5
                group_expiration:
                  description: expiration period in days
                  type: integer
                  example: 180
      responses:
        "201":
          description: invitation updated
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
    delete:
      summary: delete an invitation
      description: delete an invitation for a group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
        - in: path
          name: userUuid
          description: user uuid of invitation to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: invitation deleted
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
...
  "/groups/api/v1/self/invitations/{groupName}":
    delete:
      summary: reject an invitation to a group
      description: reject an invitation to a group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
      responses:
        "201":
          description: rejected invitation
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
    post:
      summary: join a group
      description: join a group
      parameters:
        - in: path
          name: groupName
          description: pass in the group name to interact with
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        "200":
          description: joining status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, joined]
                    example: joined
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "403":
          description: operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "406":
          description: not acceptable
...
servers:
  - url: http://localhost:8085
components:
  schemas:
    GenericError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Group name exists
    Group:
      type: object
      required:
        - id
        - name
        - type
        - description
      properties:
        name:
          type: string
          example: dinopark-test
        type:
          type: string
          enum:
            - closed
            - reviewed
            - open
          example: closed
        description:
          type: string
          example: some fancy description in markdown
    GroupWithTerms:
      type: object
      required:
        - name
        - type
        - description
      properties:
        name:
          type: string
          example: dinopark-test
        type:
          type: string
          enum:
            - closed
            - reviewed
            - open
          example: closed
        description:
          type: string
          example: some fancy description in markdown
        expiration:
          type: integer
          example: 90
        created:
          type: string
          format: datetime
          example: 2014-11-28T12:45:59Z
        terms:
          type: boolean
          example: true
    NewGroup:
      type: object
      required:
        - type
        - description
      properties:
        name:
          type: string
          example: dinos
        type:
          type: string
          enum:
            - closed
            - reviewed
            - open
          example: closed
        description:
          type: string
          example: some fancy description in markdown
        group_expiration:
          type: integer
          description: period in days after which a membership expires (0 for never)
          example: 90
    GroupUpdate:
      type: object
      properties:
        type:
          type: string
          enum:
            - closed
            - reviewed
            - open
          example: reviewed
        description:
          type: string
          example: another description in markdown
        group_expiration:
          type: integer
          description: period in days after which a membership expires (0 for never)
          example: 90
...
    DisplayMember:
      type: object
      properties:
        user_uuid:
          type: string
          format: uuid
          example: 34a33a8e-d035-4051-acac-497a79235e73
        picture:
          type: string
          nullable: true
          example: https://picture.api.dev.sso.allizom.org/avatar/get/id/MzI0ZjAwMzA5Yzc1ZGYxYmIzYjdjMGE1NWZmMDc3OGY3NGJjYjAzMjc0NDVhZmI2YWViODBmMzIzOGNhYzM3NyNzdGFmZiMxNTcyODYyNzM3.png
        first_name:
          type: string
          nullable: true
          example: Hans
        last_name:
          type: string
          nullable: true
          example: Knall
        username:
          type: string
          example: hans
        email:
          type: string
          format: email
          nullable: true
          example: hans@knall.org
        isStaff:
          type: boolean
          example: true
        since:
          type: string
          format: datetime
          nullable: true
          example: 2014-11-28T12:45:59Z
        expiration:
          type: string
          format: datetime
          nullable: true
          example: 2014-11-28T12:45:59Z
        role:
          type: string
          enum:
            - Admin
            - Curator
            - Member
        added_by:
          $ref: "#/components/schemas/AddedBy"
...
    Invitation:
      type: object
      properties:
        user_uuid:
          type: string
          format: uuid
          example: 34a33a8e-d035-4051-acac-497a79235e73
        picture:
          type: string
          nullable: true
          example: https://picture.api.dev.sso.allizom.org/avatar/get/id/MzI0ZjAwMzA5Yzc1ZGYxYmIzYjdjMGE1NWZmMDc3OGY3NGJjYjAzMjc0NDVhZmI2YWViODBmMzIzOGNhYzM3NyNzdGFmZiMxNTcyODYyNzM3.png
        first_name:
          type: string
          nullable: true
          example: Hans
        last_name:
          type: string
          nullable: true
          example: Knall
        username:
          type: string
          example: hans
        email:
          type: string
          format: email
          nullable: true
          example: hans@knall.org
        isStaff:
          type: boolean
          example: true
        invitation_expiration:
          type: string
          format: datetime
          nullable: true
          example: 2014-11-28T12:45:59Z
        group_expiration:
          type: integer
          nullable: true
          example: 90
          description: days after which membership will expire
        added_by:
          $ref: "#/components/schemas/AddedBy"
    AddedBy:
      type: object
      properties:
        user_uuid:
          type: string
          example: 34a33a8e-d035-4051-acac-497a79235e73
        first_name:
          type: string
          nullable: true
          example: Hans
        last_name:
          type: string
          nullable: true
          example: Knall
        username:
          type: string
          nullable: true
          example: hans
        email:
          type: string
          format: email
          nullable: true
          example: hans@knall.org
    GroupDetails:
      type: object
      properties:
        curator:
          type: boolean
          example: true
        group:
          $ref: "#/components/schemas/GroupWithTerms"
        member_count:
          type: integer
          example: 50
        invitation_count:
          nullable: true
          type: integer
          example: 2
        renewal_count:
          nullable: true
          type: integer
          example: 3